[development/gfx_multi_engine_cleanup] fixing all compiler warnings

+++ source/2_platform/rex_directx/include/rex_directx/resources/dx_upload_buffer.h
@@ -26,9 +26,9 @@ namespace rex
       DxUploadBuffer& operator=(DxUploadBuffer&&) = delete;
 
       // Write data on cpu side, it returns the offset into the upload buffer where data was written to
-      s32 write_buffer_data_from_cpu(const void* data, s64 size, s32 alignment = 1);
+      s64 write_buffer_data_from_cpu(const void* data, s64 size, s32 alignment = 1);
       // Write data on cpu side, it returns the offset into the upload buffer where data was written to
-      s32 write_texture_data_from_cpu(const void* data, s32 width, s32 height, TextureFormat format);
+      s64 write_texture_data_from_cpu(const void* data, s32 width, s32 height, TextureFormat format);
 
       ID3D12Resource* dx_object();
